refactor(frontend): extract helpers from checkout, ProductForm, ayuda (#244)
Batch 11: Extract schemas, types, constants, and FAQs into co-located helper files.
- checkout/page.tsx: extract Zod schema + CheckoutFormValues type → _helpers.ts
- ProductForm.tsx: extract Zod schema + types + interfaces → product-form-helpers.ts
- ayuda/page.tsx: extract SUPPORT_EMAIL, helpCategories, platformFaqs → _helpers.ts
